# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AI-powered CV search system with GraphRAG and Hybrid Search capabilities",
        "title": "CV Search \u0026 GraphRAG API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "host": "cv-search-production.up.railway.app",
    "basePath": "/api",
    "paths": {
...
        "/candidates/merge": {
            "post": {
                "description": "Atomically merges duplicate candidates and their graph relationships (skills, companies, education, CV files, interviews, scores) into a master candidate, then deletes the duplicate records.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "candidates"
                ],
                "summary": "Merge duplicate candidates",
                "parameters": [
                    {
                        "description": "Merge request containing master_candidate_id and duplicate_candidate_ids",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.mergeCandidatesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to merge candidates",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/cv/bulk-upload": {
            "post": {
                "description": "Upload multiple CV files at once (max configurable via MAX_BULK_FILE_COUNT)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cv"
                ],
                "summary": "Bulk upload CVs",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CV files (PDF, DOCX, TXT) — field name: files",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "207": {
                        "description": "Multi-Status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/cv/upload": {
            "post": {
                "description": "Upload a CV file (PDF/DOCX) and extract entities using LLM",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cv"
                ],
                "summary": "Upload and parse CV",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CV file (PDF or DOCX)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Candidate ID (optional)",
                        "name": "candidate_id",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/graph/skills/popular": {
            "get": {
                "description": "Get most popular skills extracted from CVs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graph"
                ],
                "summary": "Get popular skills",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limit results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/graphrag/communities/detect": {
            "post": {
                "description": "Run Leiden algorithm to detect communities in the knowledge graph",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graphrag"
                ],
                "summary": "Detect Communities",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hierarchy level (default: 0)",
                        "name": "level",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/graphrag/embeddings/generate": {
            "post": {
                "description": "Generate OpenAI embeddings for all nodes in the knowledge graph",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graphrag"
                ],
                "summary": "Generate Vector Embeddings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/graphrag/search": {
            "post": {
                "description": "Search candidates using natural language with Vector embeddings, Community detection, and LLM reasoning (Microsoft GraphRAG style)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graphrag"
                ],
                "summary": "GraphRAG Search (Vector + Community + LLM)",
                "parameters": [
                    {
                        "description": "Natural language search query",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.GraphRAGSearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/search": {
            "post": {
                "description": "Search for candidates based on criteria (name, location, skills)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "candidates"
                ],
                "summary": "Search candidates",
                "parameters": [
                    {
                        "description": "Search criteria",
                        "name": "criteria",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.Criteria"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.Candidate"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/search/hybrid": {
            "post": {
                "description": "Search candidates using multi-source retrieval with fusion and pure LLM reranking",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Hybrid Search (BM25 + Vector + Graph + LLM)",
                "parameters": [
                    {
                        "description": "Hybrid search request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.HybridSearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.HybridSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.FusedCandidateResponse": {
            "type": "object",
            "properties": {
                "bm25_score": {
                    "type": "number"
                },
                "communities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "community": {
                    "type": "string"
                },
                "community_scores": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                },
                "companies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphrag.CompanyNode"
                    }
                },
                "computed_community_id": {
                    "type": "string"
                },
                "computed_community_summary": {
                    "type": "string"
                },
                "current_position": {
                    "type": "string"
                },
                "fusion_score": {
                    "type": "number"
                },
                "graph_score": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "interviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.InterviewSummaryResponse"
                    }
                },
                "llm_reasoning": {
                    "type": "string"
                },
                "llm_score": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "person_id": {
                    "type": "string"
                },
                "rank": {
                    "type": "integer"
                },
                "seniority": {
                    "type": "string"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphrag.SkillNode"
                    }
                },
                "total_experience_years": {
                    "type": "integer"
                },
                "vector_score": {
                    "type": "number"
                }
            }
        },
        "api.GraphRAGSearchRequest": {
            "type": "object",
            "properties": {
                "query": {
                    "type": "string"
                }
            }
        },
        "api.HybridSearchRequest": {
            "type": "object",
            "properties": {
                "bm25_weight": {
                    "description": "Default: 0.3",
                    "type": "number"
                },
                "final_top_n": {
                    "description": "Max candidates to send to LLM (default: 0 = all)",
                    "type": "integer"
                },
                "graph_weight": {
                    "description": "Default: 0.3",
                    "type": "number"
                },
                "query": {
                    "type": "string"
                },
                "top_k": {
                    "description": "Per-source retrieval limit (default: 100)",
                    "type": "integer"
                },
                "vector_weight": {
                    "description": "Default: 0.4",
                    "type": "number"
                }
            }
        },
        "api.HybridSearchResponse": {
            "type": "object",
            "properties": {
                "candidates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.FusedCandidateResponse"
                    }
                },
                "config": {
                    "$ref": "#/definitions/graphrag.HybridSearchConfig"
                },
                "method": {
                    "type": "string"
                },
                "processing_time": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "total_found": {
                    "type": "integer"
                }
            }
        },
        "api.InterviewSummaryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "interview_date": {
                    "type": "string"
                },
                "interview_type": {
                    "type": "string"
                },
                "interviewer_name": {
                    "type": "string"
                },
                "outcome": {
                    "type": "string"
                },
                "team": {
                    "type": "string"
                }
            }
        },
        "api.mergeCandidatesRequest": {
            "type": "object",
            "properties": {
                "duplicate_candidate_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "master_candidate_id": {
                    "type": "integer"
                }
            }
        },
        "graphrag.CompanyNode": {
            "type": "object",
            "properties": {
                "is_current": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                }
            }
        },
        "graphrag.HybridSearchConfig": {
            "type": "object",
            "properties": {
                "bm25Weight": {
                    "description": "Default: 0.3",
                    "type": "number",
                    "format": "float64"
                },
                "communityThreshold": {
                    "description": "Auto-enable community filter at this candidate count (default: 50)",
                    "type": "integer"
                },
                "finalTopN": {
                    "description": "How many to send to LLM for reranking",
                    "type": "integer"
                },
                "graphWeight": {
                    "description": "Default: 0.3",
                    "type": "number",
                    "format": "float64"
                },
                "topK": {
                    "description": "How many candidates to retrieve from each source",
                    "type": "integer"
                },
                "useCommunityFilter": {
                    "description": "Enable community-based filtering (default: false, enabled at 50+ candidates)",
                    "type": "boolean"
                },
                "vectorWeight": {
                    "description": "Default: 0.4",
                    "type": "number",
                    "format": "float64"
                }
            }
        },
        "graphrag.SkillNode": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "proficiency": {
                    "type": "string"
                },
                "years_of_experience": {
                    "type": "integer"
                }
            }
        },
        "storage.Candidate": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "experience": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "resume_downloaded_at": {
                    "type": "string"
                },
                "resume_file_path": {
                    "type": "string"
                },
                "resume_url": {
                    "type": "string"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "storage.Criteria": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}